[Orm] Moving Manager classes to a new Manager namespace for better organization.
