+++ src/Valkyrja/Application/Env.php
@@ -58,8 +58,8 @@ use Valkyrja\Log\Contract\Logger;
 use Valkyrja\Log\PsrLogger;
 use Valkyrja\Mail\Contract\Mailer;
 use Valkyrja\Mail\MailgunMailer;
-use Valkyrja\Orm\Contract\Manager;
-use Valkyrja\Orm\MysqlManager;
+use Valkyrja\Orm\Manager\Contract\Manager;
+use Valkyrja\Orm\Manager\MysqlManager;
 use Valkyrja\Session\Manager\Contract\Session;
 use Valkyrja\Session\Manager\PhpSession;
 use Valkyrja\Sms\Messenger\Contract\Messenger;
