+++ src/Valkyrja/Orm/Manager/PgsqlManager.php
@@ -11,7 +11,7 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace Valkyrja\Orm;
+namespace Valkyrja\Orm\Manager;
 
 use Override;
 use Valkyrja\Orm\Exception\RuntimeException;
