+++ tests/Unit/Auth/Provider/ServiceProviderTest.php
@@ -30,7 +30,7 @@ use Valkyrja\Auth\TokenAuthenticator;
 use Valkyrja\Crypt\Contract\Crypt;
 use Valkyrja\Http\Message\Request\Contract\ServerRequest;
 use Valkyrja\Jwt\Contract\Jwt;
-use Valkyrja\Orm\Contract\Manager;
+use Valkyrja\Orm\Manager\Contract\Manager;
 use Valkyrja\Session\Manager\Contract\Session;
 use Valkyrja\Tests\Unit\Container\Provider\ServiceProviderTestCase;
