+++ tests/Unit/Orm/Provider/ServiceProviderTest.php
@@ -16,15 +16,15 @@ namespace Valkyrja\Tests\Unit\Orm\Provider;
 use PDO;
 use PHPUnit\Framework\MockObject\Exception;
 use Valkyrja\Container\Contract\Container;
-use Valkyrja\Orm\Contract\Manager as Contract;
 use Valkyrja\Orm\Entity\Entity;
-use Valkyrja\Orm\InMemoryManager;
-use Valkyrja\Orm\MysqlManager;
-use Valkyrja\Orm\NullManager;
-use Valkyrja\Orm\PgsqlManager;
+use Valkyrja\Orm\Manager\Contract\Manager as Contract;
+use Valkyrja\Orm\Manager\InMemoryManager;
+use Valkyrja\Orm\Manager\MysqlManager;
+use Valkyrja\Orm\Manager\NullManager;
+use Valkyrja\Orm\Manager\PgsqlManager;
+use Valkyrja\Orm\Manager\SqliteManager;
 use Valkyrja\Orm\Provider\ServiceProvider;
 use Valkyrja\Orm\Repository\Repository;
-use Valkyrja\Orm\SqliteManager;
 use Valkyrja\Tests\Classes\Orm\PdoClass;
 use Valkyrja\Tests\Unit\Container\Provider\ServiceProviderTestCase;
